+++ MLEM/Misc/EnumHelper.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using Microsoft.Xna.Framework.Input;
 
 namespace MLEM.Misc {
@@ -12,19 +10,20 @@ namespace MLEM.Misc {
         /// <summary>
         /// All values of the <see cref="Buttons"/> enum.
         /// </summary>
-        public static readonly Buttons[] Buttons = EnumHelper.GetValues<Buttons>().ToArray();
+        public static readonly Buttons[] Buttons = EnumHelper.GetValues<Buttons>();
         /// <summary>
         /// All values of the <see cref="Keys"/> enum.
         /// </summary>
-        public static readonly Keys[] Keys = EnumHelper.GetValues<Keys>().ToArray();
+        public static readonly Keys[] Keys = EnumHelper.GetValues<Keys>();
 
         /// <summary>
-        /// Returns all of the values of the given enum type.
+        /// Returns an array containing all of the values of the given enum type.
+        /// Note that this method is a version-independent equivalent of .NET 5's <c>Enum.GetValues&lt;TEnum&gt;</c>.
         /// </summary>
         /// <typeparam name="T">The type whose enum to get</typeparam>
         /// <returns>An enumerable of the values of the enum, in declaration order.</returns>
-        public static IEnumerable<T> GetValues<T>() {
-            return Enum.GetValues(typeof(T)).Cast<T>();
+        public static T[] GetValues<T>() {
+            return (T[]) Enum.GetValues(typeof(T));
         }
 
     }